refactor: 경로 수정 및 코드 정리
